reorganize pdfs
