+++ src/utils/openalex.py
@@ -4,39 +4,9 @@ import math
 from paper.exceptions import DOINotFoundError
 from paper.utils import get_pdf_from_url
 from researchhub.settings import OPENALEX_KEY
-from utils.aws import lambda_compress_and_linearize_pdf, upload_to_s3
-from utils.http import check_url_contains_pdf
+from utils.aws import download_pdf
 from utils.parsers import rebuild_sentence_from_inverted_index
 from utils.retryable_requests import retryable_requests_session
-from utils.sentry import log_error
-
-
-def download_pdf(url):
-    pdf_url_contains_pdf = check_url_contains_pdf(url)
-
-    if pdf_url_contains_pdf:
-        pdf_url = url
-        try:
-            pdf = get_pdf_from_url(pdf_url)
-            pdf.content_type = "application/pdf"
-            filename = pdf_url.split("/").pop()
-            if not filename.endswith(".pdf"):
-                filename += ".pdf"
-            pdf.name = filename
-            today = datetime.datetime.now()
-            year = today.year
-            month = today.month
-            day = today.day
-            folder = f"uploads/citation_entry/attachment/{year}/{month}/{day}"
-            s3_data = upload_to_s3(pdf, folder)
-            lambda_compress_and_linearize_pdf(s3_data["path"], s3_data["filename"])
-            return {"url": s3_data["url"], "signed_url": s3_data["full_url"]}
-        except Exception as e:
-            print(e)
-            log_error(e)
-            return None
-
-    return None
 
 
 class OpenAlex:
